parser/htmlparser/tests/mochitest/test_bug568470.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=568470
-->
<head>
  <title>Test for Bug 568470</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=568470">Mozilla Bug 568470</a>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 568470 **/
SimpleTest.waitForExplicitFinish();

// assuming the test runs in less than a year...
var time = new Date().getTime() + 1000 * 60 * 60 * 24 * 365;

var interval = setInterval(function() {
  var iframe = document.getElementsByTagName("iframe")[0];
  if (iframe) {
    var doc = iframe.contentDocument;
    if (doc) {
      if (doc.getElementById("flushable")) {
        time = new Date();
        clearInterval(interval);
      }
    }
  }
}, 25);

function finish() {
  clearInterval(interval);
  var elapsed = new Date().getTime() - time;
  ok(elapsed > 350,
     "Content flush time and parse end time not enough apart.");
  SimpleTest.finish();
}

</script>
</pre>
<div id="content" style="display: none">
  <iframe onload="finish();" src="file_bug568470.sjs"></iframe>
</div>
</body>
</html>
